backend/pyproject.toml

[project]
name = "summer-25-project"
version = "2020.0.0"
dependencies = [
    "pytz",
    "dotenv",
    "polygon-api-client",
    "matplotlib",
    "flask-restful",
    "flask",
    "flask_cors",
    "pandas",
]
requires-python = ">=3.9"
authors = [
    { name = "Anshika Singh" },
    { name = "Colin Berry" },
    { name = "Dylan Myers" },
]

[tool.setuptools]
package-dir = { "" = "." }
